+++ src/cache.hh
@@ -105,7 +105,7 @@ class CacheTree {
     inline tracking_vector<bool, DataStruct::Tree> opt_predictions() const;
 
     inline size_t num_nodes() const;
-    inline size_t num_nodes(size_t thread_id);
+    inline size_t num_nodes(unsigned short thread_id);
     inline size_t num_evaluated() const;
     inline size_t num_threads() const;
     inline rule_t rule(unsigned short idx) const;
@@ -131,17 +131,17 @@ class CacheTree {
 
     inline std::vector<unsigned short> rule_perm();
     void insert_root();
-    void insert(Node* node, size_t thread_id);
+    void insert(Node* node, unsigned short thread_id);
     void prune_up(Node* node);
-    void garbage_collect(size_t thread_id);
+    void garbage_collect(unsigned short thread_id);
     void print_tree();
     void close_print_file();
     void open_print_file(size_t thread_num, size_t num_threads);
     void play_with_rules();
     Node* check_prefix(tracking_vector<unsigned short, DataStruct::Tree>& prefix);
 
-    inline void lock(size_t thread_id);
-    inline void unlock(size_t thread_id);
+    inline void lock(unsigned short thread_id);
+    inline void unlock(unsigned short thread_id);
 
   protected:
     std::ofstream t_;
@@ -167,7 +167,7 @@ class CacheTree {
     rule_t *minority_;
 
     char const *type_;
-    void gc_helper(Node* node, size_t thread_id);
+    void gc_helper(Node* node, unsigned short thread_id);
     std::vector<std::mutex> tree_lks_;
 };
 
@@ -283,7 +283,7 @@ inline size_t CacheTree::num_nodes() const {
     return num_nodes_;
 }
 
-inline size_t CacheTree::num_nodes(size_t thread_id) {
+inline size_t CacheTree::num_nodes(unsigned short thread_id) {
     std::vector<unsigned short> range = get_subrange(thread_id);
     size_t ret = 1;
     for (std::vector<unsigned short>::iterator it = range.begin(); it != range.end(); ++it) {
@@ -428,13 +428,13 @@ inline void CacheTree::decrement_num_nodes() {
     logger->setTreeNumNodes(num_nodes_);
 }
 
-inline void CacheTree::lock(size_t thread_id) {
+inline void CacheTree::lock(unsigned short thread_id) {
   tree_lks_[thread_id].lock();
 }
 
-inline void CacheTree::unlock(size_t thread_id) {
+inline void CacheTree::unlock(unsigned short thread_id) {
   tree_lks_[thread_id].unlock();
 }
 
 void delete_interior(CacheTree* tree, Node* node, bool destructive, bool update_remaining_state_space);
-void delete_subtree(CacheTree* tree, Node* node, bool destructive, bool update_remaining_state_space, size_t thread_id);
+void delete_subtree(CacheTree* tree, Node* node, bool destructive, bool update_remaining_state_space, unsigned short thread_id);
