+++ src/pmap.hh
@@ -97,7 +97,7 @@ class PermutationMap {
                              size_t nrules, bool prediction, bool default_prediction, double lower_bound,
                              double objective, Node* parent, int num_not_captured, int nsamples, int len_prefix,
                              double c, double equivalent_minority, CacheTree* tree, VECTOR not_captured,
-                             tracking_vector<unsigned short, DataStruct::Tree> parent_prefix, size_t thread_id) { return NULL; }
+                             tracking_vector<unsigned short, DataStruct::Tree> parent_prefix, unsigned short thread_id) { return NULL; }
         Node* check_permutation_bound (unsigned short new_rule,
                              size_t nrules, bool prediction, bool default_prediction, double lower_bound,
                              double objective, Node* parent, int num_not_captured, int nsamples, int len_prefix,
@@ -115,7 +115,7 @@ class PrefixPermutationMap : public PermutationMap {
             bool default_prediction, double lower_bound, double objective, Node* parent, 
             int num_not_captured, int nsamples, int len_prefix, double c, double equivalent_minority,
             CacheTree* tree, VECTOR not_captured, tracking_vector<unsigned short, 
-            DataStruct::Tree> parent_prefix, size_t thread_id) override;
+            DataStruct::Tree> parent_prefix, unsigned short thread_id) override;
 	private:
 		PrefixMap* pmap;
         std::mutex map_lk;
@@ -133,7 +133,7 @@ class CapturedPermutationMap : public PermutationMap {
         Node* insert(unsigned short new_rule, size_t nrules, bool prediction, bool default_prediction, 
                 double lower_bound, double objective, Node* parent, int num_not_captured, int nsamples, 
                 int len_prefix, double c, double equivalent_minority, CacheTree* tree, VECTOR not_captured,
-                 tracking_vector<unsigned short, DataStruct::Tree> parent_prefix, size_t thread_id) override;
+                 tracking_vector<unsigned short, DataStruct::Tree> parent_prefix, unsigned short thread_id) override;
 	private:
 		CapturedMap* pmap;
         std::mutex key_lk;
@@ -146,7 +146,7 @@ class NullPermutationMap : public PermutationMap  {
         Node* insert (unsigned short new_rule, size_t nrules, bool prediction, bool default_prediction, double lower_bound,
                         double objective, Node* parent, int num_not_captured, int nsamples, int len_prefix,
                         double c, double equivalent_minority, CacheTree* tree, VECTOR not_captured,
-                        tracking_vector<unsigned short, DataStruct::Tree> parent_prefix, size_t thread_id) override {
+                        tracking_vector<unsigned short, DataStruct::Tree> parent_prefix, unsigned short thread_id) override {
             return tree->construct_node(new_rule, nrules, prediction, default_prediction,
                                             lower_bound, objective, parent,
                                             num_not_captured, nsamples, len_prefix, c, equivalent_minority);
